Enable strict compiler flags with a configure flag
When making a distcheck it's usually a good idea to enable very strict
compiler flags, like -Werror, to catch stuff that slipped through the
development phase.

This patch adds a --enable-maintainer-flags command line switch to the
configure script, which enables a set of strict compiler flags. The default
is not to use them unless explicitly activated.

In case of distcheck, this switch is activated when launching the configure
script from within the distcheck build directory.