Don't override XAUTHORITY until after we run Xephyr
Don't override the setting of XAUTHORITY when running Xephyr,
since then we may not be able to connect to the host display.
Instead pass in the local auth file via the -auth command
line argument.

An alternate fix would be to use 'xauth merge' to merge the
existing XAUTHORITY file into our local file.

svn path=/trunk/; revision=29